+++ android/src/main/java/com/reactnativestripesdk/PaymentSheetManager.kt
@@ -239,6 +239,16 @@ class PaymentSheetManager(
       return
     }
 
+    if (checkoutSessionKey != null) {
+      promise.resolve(
+        createError(
+          ErrorType.Failed.toString(),
+          "PaymentSheet with checkout is not supported. Use customFlow: true instead.",
+        ),
+      )
+      return
+    }
+
     lastConfigureWasCustomFlow = false
     if (paymentSheet == null) {
       initPaymentSheet(args, promise)
@@ -437,10 +447,7 @@ class PaymentSheetManager(
   override fun onPresent() {
     keepJsAwake = KeepJsAwakeTask(context).apply { start() }
     if (lastConfigureWasCustomFlow == false) {
-      val checkout = getCheckoutInstanceOrResolveError(promise)
-      if (checkout != null) {
-        paymentSheet?.presentWithCheckout(checkout, paymentSheetConfiguration)
-      } else if (!paymentIntentClientSecret.isNullOrEmpty()) {
+      if (!paymentIntentClientSecret.isNullOrEmpty()) {
         paymentSheet?.presentWithPaymentIntent(
           paymentIntentClientSecret!!,
           paymentSheetConfiguration,