+++ ios/StripeSdkImpl+PaymentSheet.swift
@@ -196,8 +196,7 @@ extension StripeSdkImpl {
                     handlePaymentSheetFlowControllerResult(result: result, stripeSdk: self)
                 }
             } else {
-                self.paymentSheet = PaymentSheet(checkout: checkout, configuration: configuration)
-                resolve([])
+                resolve(Errors.createError(ErrorType.Failed, "PaymentSheet with checkout is not supported. Use customFlow: true instead."))
             }
         } else if let paymentIntentClientSecret = params["paymentIntentClientSecret"] as? String {
             if !Errors.isPIClientSecretValid(clientSecret: paymentIntentClientSecret) {