+++ src/specs/NativeEmbeddedPaymentElement.ts
@@ -20,19 +20,10 @@ export interface NativeCommands {
     viewRef: React.ElementRef<HostComponent<NativeProps>>,
     intentConfigurationJson: string
   ) => void;
-  updateWithCheckout: (
-    viewRef: React.ElementRef<HostComponent<NativeProps>>,
-    sessionKey: string
-  ) => void;
 }
 
 export const Commands: NativeCommands = codegenNativeCommands<NativeCommands>({
-  supportedCommands: [
-    'confirm',
-    'clearPaymentOption',
-    'update',
-    'updateWithCheckout',
-  ],
+  supportedCommands: ['confirm', 'clearPaymentOption', 'update'],
 });
 
 type ComponentType = HostComponent<NativeProps>;