+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 **Changes**
+* Updated Stripe iOS SDK from 25.17.0 to 26.0.0.
+* Updated Stripe Android SDK from 23.10.1 to 23.11.0.
 * [Breaking] Updated Stripe Issuing Android Push Provisioning SDK from 1.1.0+ to 1.3.0+.
 * [Breaking] Updated Google Wallet Android Push Provisioning to the Unified Push Provisioning flow. This enables issued cards to be provisioned onto wearable devices via a mobile phone. `canAddCardToWallet` now requires `cardBrand`, and now checks if the card can be provisioned onto the mobile device or any connected wearable devices. See [documentation](https://docs.stripe.com/issuing/cards/digital-wallets?platform=react-native#push-provisioning) for integration guide.